pyproject.toml
[project]
name = "itacambira"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf>=1.7.2",
    "jinja2-simple-tags>=0.6.1",
    "jinja2>=3.1.4",
    "pyperclip>=1.9.0",
    "pywebview>=5.3.2",
]

[dependency-groups]
dev = [
    "djlint>=1.36.1",
    "isort>=5.13.2",
    "pyinstaller>=6.11.1",
    "ruff>=0.7.4",
    "watchfiles>=0.24.0",
]


[tool.ruff.format]
quote-style = 'single'


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PT']

[tool.djlint]
profile = "jinja"
